[#91] 피드 좋아요 Redis 올리기
- RedisTemplate 타입 변경

+++ src/main/java/com/social/instagram/config/redis/FeedsNiceRedisConfig.java
@@ -24,8 +24,8 @@ public class FeedsNiceRedisConfig {
     }
 
     @Bean
-    public RedisTemplate<Integer, Integer> feedsNiceRedisTemplate() {
-        RedisTemplate<Integer, Integer> feedsNiceRedisTemplate = new RedisTemplate<>();
+    public RedisTemplate<Long, Long> feedsNiceRedisTemplate() {
+        RedisTemplate<Long, Long> feedsNiceRedisTemplate = new RedisTemplate<>();
         feedsNiceRedisTemplate.setConnectionFactory(feedsNiceRedisConnectionFactory());
         feedsNiceRedisTemplate.setKeySerializer(new GenericJackson2JsonRedisSerializer());
         feedsNiceRedisTemplate.setValueSerializer(new GenericJackson2JsonRedisSerializer());
